#2 -change: Moved functions to environment.py
