+++ funl/utils/globalvars.py
@@ -23,25 +23,3 @@ from .function_definition import FunctionDefinition
 VERSION = "a-2"
 environment: list[list[FunctionDefinition]] = [[]]
 
-
-def get_function_from_name(name: str) -> FunctionDefinition | None:
-    for environment_function in environment:
-        if environment_function.name == name:
-            return environment_function
-    return None
-
-
-def get_function_index_from_name(name: str) -> int | None:
-    for i in range(0, len(environment)):
-        if environment[i].name == name:
-            return i
-    return None
-
-
-def append_or_update(function_definition: FunctionDefinition) -> None:
-    result = get_function_index_from_name(function_definition.name)
-
-    if result is not None:
-        environment[result] = function_definition
-    else:
-        environment.append(function_definition)
